Refactor image translator into package
