+++ README.md
@@ -1,9 +1,9 @@
 В этот репозиторий будут выкладываться "скетчи", зачастую возможно не хорошо написанные и/или имеющие какую-то хорошую идею.
 * Image Translator ("Переводчик изображений") - с помощью WINAPI делает скриншот выбранной области экрана, распознает английский текст с помощью OCR, затем переводит его на русский и выводит переведенный текст на экран.
 
-  Файл `image_translator.py` предполагает, что Tesseract установлен по пути
+  Скрипт `image_translator/translator_app.py` предполагает, что Tesseract установлен по пути
   `C:\Program Files\Tesseract-OCR\tesseract.exe`. Если путь к исполняемому
-  файлу другой, отредактируйте переменную `tesseract_cmd` в коде.
+  файлу другой, отредактируйте функцию `configure_tesseract` в модуле `image_translator.utils`.
 
 ## Установка зависимостей
