+++ dev-image_translator/README.md
@@ -4,7 +4,7 @@
 
 ## Как это работает
 
-1. Запускается скрипт `image_translator.py`.
+1. Запускается скрипт `translator_app.py` из подпакета `image_translator`.
 2. Пользователь нажимает и удерживает комбинацию `CTRL+ALT`, перемещая курсор для выделения прямоугольной области на экране.
 3. Выбранная область сохраняется как изображение и передаётся в OCR (pytesseract).
 4. Распознанный английский текст переводится на русский с помощью библиотеки `translate`.
@@ -29,13 +29,13 @@ pip install -r requirements.txt
 ## Запуск
 
 ```bash
-python image_translator.py
+python -m image_translator.translator_app
 ```
 
 Для вывода каждого снятого скриншота используйте режим отладки:
 
 ```bash
-python image_translator.py --debug
+python -m image_translator.translator_app --debug
 ```
 
 После запуска одновременно нажмите `CTRL` и `ALT` и, не отпуская их, выделите область с текстом. Когда клавиши будут отпущены, скрипт распознает текст, переведёт его и выведет результат поверх остальных окон.
