+++ image_translator/__init__.py
@@ -1,0 +1,5 @@
+"""Simple image translator package."""
+
+from .translator_app import ImageTranslatorApp, main
+
+__all__ = ["ImageTranslatorApp", "main"]
